requirement.txt
streamlit
groq
selenium
python-dotenv
webdriver-manager